+++ gatling-charts-highcharts/src/main/scala/com/excilys/ebi/gatling/highcharts/series/ColumnSeries.scala
@@ -7,7 +7,7 @@ package com.excilys.ebi.gatling.highcharts.series
 
 import com.excilys.ebi.gatling.charts.series.Series
 
-class ColumnSeries(name: String, data: Seq[(String, Long)], colors: List[String]) extends Series[String, Long](name, data, colors) {
+class ColumnSeries(name: String, data: Seq[(String, Int)], colors: List[String]) extends Series[String, Int](name, data, colors) {
 	val getXValues = data.map(_._1)
 	val getYValues = data.map(_._2)
 }
