+++ gatling-charts-highcharts/src/main/scala/com/excilys/ebi/gatling/highcharts/series/StackedColumnSeries.scala
@@ -5,7 +5,7 @@
  */
 package com.excilys.ebi.gatling.highcharts.series
 
-class StackedColumnSeries(name: String, data: Seq[(String, Long)], color: String) extends ColumnSeries(name, data, List(color)) {
+class StackedColumnSeries(name: String, data: Seq[(String, Int)], color: String) extends ColumnSeries(name, data, List(color)) {
 
 	def elements: Seq[String] = data.map { case (_, count) => count.toString }
 }
